fix review 2 сделал метод для появления Алерта универсальным
